requirements.txt
altgraph==0.17.2
blinker==1.9.0
cffi==1.15.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
Flask-Login==0.6.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
Werkzeug==3.1.3